Progress moving dao tests around
