+++ packages/aragon-cli/test/dao/dao-new.test.js
@@ -2,10 +2,10 @@ import test from 'ava'
 import execa from 'execa'
 import { normalizeOutput } from '../util'
 
-const daoAddressRegex = /Created DAO\: (.*)$/
-const daoIdAndAddressAddressRegex = /Created DAO\: (.*) at (.*)$/
+const daoAddressRegex = /Created DAO: (.*)$/
+const daoIdAndAddressAddressRegex = /Created DAO: (.*) at (.*)$/
 
-test.skip('creates a new DAO', async t => {
+test('creates a new DAO', async t => {
   t.plan(2)
 
   const result = await execa('aragon', ['dao', 'new', '--debug'])
@@ -17,7 +17,7 @@ test.skip('creates a new DAO', async t => {
   t.snapshot(resultSnapshot)
 })
 
-test.skip('assigns an Aragon Id with the "--aragon-id" param', async t => {
+test('assigns an Aragon Id with the "--aragon-id" param', async t => {
   t.plan(3)
 
   const result = await execa('aragon', [
