+++ packages/toolkit/test/test-helpers/getLocalWeb3.js
@@ -1,0 +1,12 @@
+import Web3 from 'web3'
+
+const getLocalWeb3 = async () => {
+  const web3 = new Web3(
+    new Web3.providers.WebsocketProvider(`ws://localhost:8545`)
+  )
+  const connected = await web3.eth.net.isListening()
+  if (!connected) throw new Error('Web3 connection failed')
+  return web3
+}
+
+export default getLocalWeb3
