+++ packages/toolkit/test/test-helpers/index.js
@@ -1,0 +1,4 @@
+import getLocalWeb3 from './getLocalWeb3'
+import isAddress from './isAddress'
+
+export { getLocalWeb3, isAddress }
