+++ packages/toolkit/test/test-helpers/isAddress.js
@@ -1,0 +1,5 @@
+import Web3 from 'web3'
+
+const isAddress = Web3.utils.isAddress
+
+export default isAddress
